OptParse

Description

A full-featured command line arguments parser. Features:

Parses both short options (-i) and long options (--ignore)

Allows short options to be combined with other short options (ls -lh rather than ls -l -h) and with arguments (foocmd -oarg rather than foocmd -o arg and even foocmd -ioarg assuming that -i is an option that takes no arguments)

Allows long option to be combined with its argument with an equal sign (GNU style, like head --lines=20). However this is optional; having the long option and the argument in separate words is also permitted (like head --lines 20).

Parses "commands". Examples include git clone or cvs up, where clone and up are the commands, respectively. Each command might take different options and handle positional arguments differently.

Both commands and options can be abbreviated to the shortest unambiguous abbreviation. For example, instead of --ignore-case you could use --igno or even --i if either of those is unambiguous.

Options, option arguments, commands, and positional arguments are all parsed using user-provided functions. For example, you can use this to change option arguments to different datatypes (an Int, say) or to make sure there is a particular number of positional arguments. You could make sure some options only appear on the command line once, or you can make implement an option that toggles an option on and off.

Full error reporting using Either and a user-defined error object. Parsing stops when an error is found, and that error is reported. The user can define her own errors, which can be used if an error is encountered when parsing options, option arguments, commands, or positional arguments.

Positional arguments and options can be freely interspersed.

Recognizes -- by itself and then stops parsing options. All subsequent words are then treated as positional arguments, even if they begin with dashes.

Options can take zero, one, or two arguments.

Non-features / disadvantages:

Complicated.

Some parsers, like System.Console.GetOpt, automatically generate help messages; this does not.

Options must take a set number of arguments. Unlike System.Console.GetOpt, you cannot specify that an option takes an optional argument.

Short options (those specified with a single dash) must be only one character long. So you will not be able to use this to write a clone of the Unix find(1) command.

Although short options can be combined (e.g. -isn assuming -i and -s take no option arguments and that -n is an option) the first option must be preceded with a dash. That is, this does not support the style of options you find in BSD-like ps(1) commands, for example.

Positional arguments and options may always be freely interspersed; this cannot be turned off.

Documentation

class ParseErr a where

Instances of the ParseErr class can be used with the parse function to report errors. In addition, ParseErr can be used to report custom errors. For example, one of the argument processing functions in ArgDesc can refurn a Left err to indicate that the user passed an incorrect value. This will terminate all further processing and report the error. Though options are typically preceded with one or two dashes, none of the arguments passed to the ParseErr functions are passed with leading dashes.

Methods

badLongOpt

Arguments

:: String
The bad input

-> a

A long option string was not recognized.

ambiguousLongOpt

Arguments

:: String
The string the user passed

-> [String]
Possible matches

-> a

A long option string was short but has more than one match. For instance the user passed the string --he as an option, but the program has options for both --hello and --head.

badShortOpt

Arguments

:: Char
Unrecognized character

-> a

A short option character has no recognized matches.

badEqualsOpt

Arguments

:: String
The option name

-> String
The option argument the user passed

-> a

Long options can be passed GNU style, where the option and the option argument appear in the same word, as in --number=4. This works when the option has one or two arguments. When an option argument is passed for an option that does not take any arguments, this function is called.

insufficientArgs

Arguments

:: Either String Char
The option name (short or long)

-> Int
The number of arguments expected

-> [String]
Number of arguments actually received

-> a

The user did not give enough option arguments.

longOptWithoutName :: a

Long option was passed, and it has an equal sign and an option argument, but no name (e.g. --=yes)

noMatchingCmd

Arguments

:: String
name of unrecognized command

-> a

An unrecognized command name was used.

ambiguousCmd

Arguments

:: String
The string the user passed

-> [String]
Possible matches

-> a

A command name matches more than one possible command name, for instance the user passed fin but there are commands named find and finagle.

noCmd :: a

The user did not provide any command at all.

posArgsNotAllowed

Arguments

:: String
Positional argument the user gave

-> a

The user passed a positional argument where one is not allowed (e.g. for a command for which positional arguments are not allowed).

data PosDesc opts posargs err

Describes a positional argument. For example, in the command pantry find -i name hello, the positional arguments are name and hello (-i is an option, not a positional argument).

Constructors

NoPosArgs posargs
Use if this command does not accept any positional arguments. However you still must give a function that takes no arguments; this function is called when and will be suppled as one of the elements of the tuple that parse returns. (Possible change: have parse return a Nothing instead in cases like this?)

PosArgs ([(opts, String)] -> Either err posargs)
parse parses all the positional arguments after the rest of the command line is processed. You supply the function that will be called here. This way the function can see all the command line arguments rather than seeing them just one at a time. Each element of this is a tuple, with the first part of the tuple being the options that were selected at the time this positional argument appeared, and the second being the positional argument itself. This way the output of the function can depend on what the command line options were. The function should return Left err if there was some problem.

data ArgDesc opts err

Describes the arguments that an option accepts. For example, the command head takes an option named -n, which accepts a single argument for the number of lines to print. Each of these constructors takes one argument, which is a function. The first argument of the function is always the command line options as they existed before this argument is parsed. This allows each option behave differently depending upon what options preceded it. For instance, this could be used to implement toggling behavior. Next the function accepts zero or more string arguments, with each string representing a command-line argument for the option. Finally each function returns an Either. Return Left err if there was some problem with the parse, or a Right opts with the new state of the command-line options.

Constructors

Flag (opts -> Either err opts)

Single (opts -> String -> Either err opts)

Double (opts -> String -> String -> Either err opts)

data OptDesc opts err

Describes a command line option. For example, in the command line pantry find -i name Chex, a command line option is -i.

Constructors

OptDesc

Fields

optShort :: [Char]

Short option names, without the leading dash.

optLong :: [String]

Long option names, without the leading dashes.

optArgDesc :: ArgDesc opts err

Whether this option takes any arguments and, if so, how many.

data CmdDesc cmd opts posargs err

Describes a command. For instance, from the command line pantry find -i name Chex, the command is find.

Constructors

CmdDesc

Fields

cmdName :: String

The command name, such as find.

cmdF :: String -> cmd

Takes the full name of the command, returns something to indicate which command was called.

cmdOpts :: [OptDesc opts err]

A list of all the options that the command takes.

cmdPos :: PosDesc opts posargs err

How to parse positional arguments for this command.

parse

Arguments

:: (ParseErr err, Error err)

=> [OptDesc opts err]
Global options. For example, in the command pantry -i find name pretzels, -i is a global option because it appears before the command.

-> [CmdDesc cmd opts posargs err]
All command descriptions

-> opts
Default options. When the command line is parsed, each option receives the options that have already been parsed. The first option will receive this item.

-> [String]
What to parse

-> Either err (cmd, opts, posargs)
Left if an error occurred; Right if everything succeeded. The tuple has the command that was seen, the final state of the options item, and any positional arguments.

Parse a command line. Presumably you got it from System.Environment.getArgs. Do not include the program name as the first string to parse (consistent with System.Environment.getArgs).

type CharOpts opts err = Map Char (ArgDesc opts err)

An ArgDesc describes an argument, but the ArgDesc itself does not specify the name of the short or long option that is used on the command line to trigger the argument. For example, a program might have a flag option (one that does not take any option arguments). The flag option might be invoked using -f or with --flag. To parse these flags, add appropriate key-value pairs to the CharOpts and StringOpts maps.

type StringOpts opts err = Map String (ArgDesc opts err)

addCharOpt :: ArgDesc opts err -> CharOpts opts err -> Char -> CharOpts opts err

Adds a single argument description to a map of CharOpts, which will be triggered by the given char.

addStringOpt :: ArgDesc opts err -> StringOpts opts err -> String -> StringOpts opts err

Adds a single argument description to a map of StringOpts.

addCharOpts :: CharOpts opts err -> OptDesc opts err -> CharOpts opts err

Adds the list of short option names found in an OptDesc to the CharOpts, with each short option name being associated with the ArgDesc given in the OptDesc.

addStringOpts :: StringOpts opts err -> OptDesc opts err -> StringOpts opts err

Adds the list of long argument names found in an OptDesc to the StringOpts, with each long argument name being associated with the ArgDesc given in the OptDesc.

addOptsToLookups :: [OptDesc opts err] -> (CharOpts opts err, StringOpts opts err)

Applies addStringOpts and addCharOpts repeatedly to a list of OptDesc items.

addCmdToLookups :: CmdDesc cmd opts posargs err -> (CharOpts opts err, StringOpts opts err)

A simple wrapper around addOptsToLookups for CmdDesc items.

data ParseState opts

Constructors

ParseState

Fields

stOpts :: opts

Options parsed so far

stPos :: [(opts, String)]

Positional arguments so far. These are not truly parsed; they are simply paired with the options that were in force when the positional argument was encountered on the command line.

stLeft :: [String]

Words remaining to be parsed

isLongOpt :: String -> Bool

isShortOpt :: String -> Bool

parseCmdDesc

Arguments

:: (ParseErr err, Error err)

=> CmdDesc cmd opts posargs err
The CmdDesc of the command to parse, which includes all the options and how to parse the positional arguments.

-> [String]
What to parse. Using the example given above, this would be

 ["-i", "name", "beans"]

-> opts
Default options. Each time a set of option arguments is parsed (which happens even for options that take no arguments) a function is called that takes the set of options as it has been modified by the command line so far. The value if this parameter is passed to the first option.

-> Either err (opts, [(opts, String)])
This will be Left err if an error occurred, or Right if everything succeeded. The first element of the tuple is the options after all have been parsed. The second is a list of positional arguments. The first element of each tuple in this list is the state of the options when the argument was encountered; the second element is the argument itself.

Parses a single command description. At this point, the command itself has already been removed from the command line. For example, if the command line is pantry find -i name beans, then the command is find. However, the line has already been parsed through find.

data AtNonOpt

What to do when the args parser encounters a non-option argument? Both of these options will properly respect a "stopper" (--). The stopper will be parsed and then the remaining arguments will be returned unparsed.

Constructors

StopParsing
Stop parsing, do not parse the non-option argument, and return the remaining arguments unparsed

ResumeParsing
Add the non-option argument to the list of remaining arguments that will be returned, then resume parsing at the next word

parseArgs

Arguments

:: (ParseErr err, Error err)

=> AtNonOpt
See documentation for AtNonOpt for details

-> CharOpts opts err
The short options

-> StringOpts opts err
The long options

-> [String]
What to parse

-> opts
Initial state of the options

-> Either err (opts, [(opts, String)])
This will be Left err if an error occurred, or Right if everything succeeded. The first element of the tuple is the options after all have been parsed. The second is a list of positional arguments. The first element of each tuple in this list is the state of the options when the argument was encountered; the second element is the argument itself.

Parses a list of options and positional arguments.

parseArgsM :: (ParseErr err, Error err) => AtNonOpt -> CharOpts opts err -> StringOpts opts err -> ErrorT err (State (ParseState opts)) ()

Monad that carries out the parsing of options and positional arguments. This function recurses until all the command line arguments are parsed (either because they are options, or because they are positional arguments).

pickParser :: (ParseErr err, Error err) => AtNonOpt -> String -> CharOpts opts err -> StringOpts opts err -> ErrorT err (State (ParseState opts)) ()

Examines, but does not change, the next word in the ParseState to be parsed. Then calls the appropriate parser in to actually parse the word. This function does not actually change or remove the next word to be parsed because the parser itself might need to examine the word (for instance, it might be a long option that also has the option argument).

isStopper :: String -> Bool

A "stopper" is a double dash and means "stop processing options; everything that follows is a positional argument."

parseStopper :: (ParseErr err, Error err) => ErrorT err (State (ParseState opts)) ()

Removes the stopper from the list of arguments remaining to be parsed, then parses all the remaining arguments as positional arguments.

parseLongOpt :: (ParseErr err, Error err) => StringOpts opts err -> ErrorT err (State (ParseState opts)) ()

Parse a long option. pickParser has already figured out that the next word to be parsed is a long option; parseLongOpt examines the next word, examines StringOpts to determine how many option arguments it takes, and parses the option arguments. This uses breakLongWord to split the long option into the option itself and any argument that may have been passed in the same word. If the long option is not found, this calls throwError.

pickLongParser

Arguments

:: (ParseErr err, Error err)

=> StringOpts opts err

-> (String, Maybe String)
The result from breakLongWord. fst is the name of the long option. snd is the option argument that was passed in the same word, or Nothing if there wasn't one. See breakLongWord for details.

-> ErrorT err (State (ParseState opts)) ()

There are three different functions to parse long flag options, long single options, and long double options; this function picks the right one.

parseLongFlag :: (ParseErr err, Error err) => (opts -> Either err opts) -> (String, Maybe String) -> ErrorT err (State (ParseState opts)) ()

parseLongSingle :: (ParseErr err, Error err) => (opts -> String -> Either err opts) -> (String, Maybe String) -> ErrorT err (State (ParseState opts)) ()

parseLongDouble :: (ParseErr err, Error err) => (opts -> String -> String -> Either err opts) -> (String, Maybe String) -> ErrorT err (State (ParseState opts)) ()

breakLongWord :: (Error err, ParseErr err) => String -> Either err (String, Maybe String)

Applied to a string that is a long option, such as --hello=yes, returns the long option name and the argument given, if any. For example, for --hello=yes, returns (hello, Just yes). For --hello, returns (hello, Nothing). Returns an error if there is no long option name (e.g. --=yes).

bestLongArgDesc :: ParseErr err => StringOpts opts err -> String -> Either err (ArgDesc opts err)

Finds the ArgDesc that is the best match for a string. If there is an exact match, use that. Otherwise, if there is exactly one option that starts with the word given, use that. Otherwise, returns an error.

parseShortOpt :: (ParseErr err, Error err) => CharOpts opts err -> ErrorT err (State (ParseState opts)) ()

With short options, multiple options can be combined into one word, and an option and its argument might be combined in the same word. parseShortOpt pops the first word off from the list of words to be parsed (pickParser already ensured that the first word is indeed a short option). parseShortOpt peels the single dash off the front of the word and then passes the word to innerParseShort for parsing. BUG - FIXME - no recursion for innerParseShort

innerParseShort

Arguments

:: (ParseErr err, Error err)

=> CharOpts opts err

-> [Char]
The word with the option - prefix omitted

-> ErrorT err (State (ParseState opts)) ()

parseShortFlag :: (ParseErr err, Error err) => (opts -> Either err opts) -> CharOpts opts err -> [Char] -> ErrorT err (State (ParseState opts)) ()

parseShortSingle :: (ParseErr err, Error err) => (opts -> String -> Either err opts) -> Char -> [Char] -> ErrorT err (State (ParseState opts)) ()

parseShortDouble :: (ParseErr err, Error err) => (opts -> String -> String -> Either err opts) -> Char -> [Char] -> ErrorT err (State (ParseState opts)) ()

parseOnePosArg :: (ParseErr err, Error err) => ErrorT err (State (ParseState opts)) ()

parseRemainingPosArgs :: (ParseErr err, Error err) => ErrorT err (State (ParseState opts)) ()

parseCmds

Arguments

:: (ParseErr err, Error err)

=> [CmdDesc cmd opts posargs err]

-> opts
Default opts

-> [String]
To parse

-> Either err (cmd, opts, posargs)

Applied to a list of command descriptions, a set of default options, and a list of strings to parse, returns either an error or the resulting command, the resulting options, and the resulting posargs.

pickCmd

Arguments

:: (ParseErr err, Error err)

=> [CmdDesc cmd opts posargs err]

-> [String]
To parse

-> Either err (CmdDesc cmd opts posargs err)

Given a list of CmdDesc and a list of command line arguments waiting to be parsed, returns the appropriate CmdDesc for the argument at the head of the list. Assumes there is at least one argument in the list; blows up if there isn't one.

throwOnLeft :: MonadError e m => Either e b -> m ()

Throws an error if applied to an Either which is Left. Otherwise, does nothing.

fromRight :: Either a b -> b

Extracts the Right value from an Either. Bombs if there is no Right value.