OptParse

Description

A full-featured command line arguments parser. Features:

Parses both short options (-i) and long options (--ignore)

Allows short options to be combined with other short options (ls -lh rather than ls -l -h) and with arguments (foocmd -oarg rather than foocmd -o arg and even foocmd -ioarg assuming that -i is an option that takes no arguments)

Allows long option to be combined with its argument with an equal sign (GNU style, like head --lines=20). However this is optional; having the long option and the argument in separate words is also permitted (like head --lines 20).

Parses "commands". Examples include git clone or cvs up, where clone and up are the commands, respectively. Each command might take different options and handle positional arguments differently.

Both commands and options can be abbreviated to the shortest unambiguous abbreviation. For example, instead of --ignore-case you could use --igno or even --i if either of those is unambiguous.

Options, option arguments, commands, and positional arguments are all parsed using user-provided functions. For example, you can use this to change option arguments to different datatypes (an Int, say) or to make sure there is a particular number of positional arguments. You could make sure some options only appear on the command line once, or you can make implement an option that toggles an option on and off.

Full error reporting using Either and a user-defined error object. Parsing stops when an error is found, and that error is reported. The user can define her own errors, which can be used if an error is encountered when parsing options, option arguments, commands, or positional arguments.

Positional arguments and options can be freely interspersed.

Recognizes -- by itself and then stops parsing options. All subsequent words are then treated as positional arguments, even if they begin with dashes.

Options can take zero, one, or two arguments.

Non-features / disadvantages:

Complicated.

Some parsers, like System.Console.GetOpt, automatically generate help messages; this does not.

Options must take a set number of arguments. Unlike System.Console.GetOpt, you cannot specify that an option takes an optional argument.

Short options (those specified with a single dash) must be only one character long. So you will not be able to use this to write a clone of the Unix find(1) command.

Although short options can be combined (e.g. -isn assuming -i and -s take no option arguments and that -n is an option) the first option must be preceded with a dash. That is, this does not support the style of options you find in BSD-like ps(1) commands, for example.

Positional arguments and options may always be freely interspersed; this cannot be turned off.

Documentation

class ParseErr a where

Instances of the ParseErr class can be used with the parse function to report errors. In addition, ParseErr can be used to report custom errors. For example, one of the argument processing functions in ArgDesc can refurn a Left err to indicate that the user passed an incorrect value. This will terminate all further processing and report the error. Though options are typically preceded with one or two dashes, none of the arguments passed to the ParseErr functions are passed with leading dashes.

Minimal complete definition: store.

Methods

store :: String -> a

The default implementations of the class functions use the store function to store error messages in the class instance.

badLongOpt

Arguments

:: String
The bad input

-> a

A long option string was not recognized.

ambiguousLongOpt

Arguments

:: String
The string the user passed

-> [String]
Possible matches

-> a

A long option string was short but has more than one match. For instance the user passed the string --he as an option, but the program has options for both --hello and --head.

badShortOpt

Arguments

:: Char
Unrecognized character

-> a

A short option character has no recognized matches.

badEqualsOpt

Arguments

:: String
The option name

-> String
The option argument the user passed

-> a

Long options can be passed GNU style, where the option and the option argument appear in the same word, as in --number=4. This works when the option has one or two arguments. When an option argument is passed for an option that does not take any arguments, this function is called.

insufficientArgs

Arguments

:: Either String Char
The option name (short or long)

-> Int
The number of arguments expected

-> [String]
Number of arguments actually received

-> a

The user did not give enough option arguments.

longOptWithoutName

Arguments

:: String
Text of the bad long option

-> a

Long option was passed, and it has an equal sign and an option argument, but no name (e.g. --=yes)

noMatchingCmd

Arguments

:: String
name of unrecognized command

-> a

An unrecognized command name was used.

ambiguousCmd

Arguments

:: String
The string the user passed

-> [String]
Possible matches

-> a

A command name matches more than one possible command name, for instance the user passed fin but there are commands named find and finagle.

noCmd :: a

The user did not provide any command at all.

posArgsNotAllowed

Arguments

:: String
Positional argument the user gave

-> a

The user passed a positional argument where one is not allowed (e.g. for a command for which positional arguments are not allowed).

data PosDesc opts posargs err

Describes a positional argument. For example, in the command pantry find -i name hello, the positional arguments are name and hello (-i is an option, not a positional argument).

Constructors

NoPosArgs posargs
Use if this command does not accept any positional arguments. However you still must give a function that takes no arguments; this function is called when and will be suppled as one of the elements of the tuple that parse returns. (Possible change: have parse return a Nothing instead in cases like this?)

PosArgs ([(opts, String)] -> Either err posargs)
parse parses all the positional arguments after the rest of the command line is processed. You supply the function that will be called here. This way the function can see all the command line arguments rather than seeing them just one at a time. Each element of this is a tuple, with the first part of the tuple being the options that were selected at the time this positional argument appeared, and the second being the positional argument itself. This way the output of the function can depend on what the command line options were. The function should return Left err if there was some problem.

data ArgDesc opts err

Describes the arguments that an option accepts. For example, the command head takes an option named -n, which accepts a single argument for the number of lines to print. Each of these constructors takes one argument, which is a function. The first argument of the function is always the command line options as they existed before this argument is parsed. This allows each option behave differently depending upon what options preceded it. For instance, this could be used to implement toggling behavior. Next the function accepts zero or more string arguments, with each string representing a command-line argument for the option. Finally each function returns an Either. Return Left err if there was some problem with the parse, or a Right opts with the new state of the command-line options.

Constructors

Flag (opts -> Either err opts)

Single (opts -> String -> Either err opts)

Double (opts -> String -> String -> Either err opts)

data OptDesc opts err

Describes a command line option. For example, in the command line pantry find -i name Chex, a command line option is -i.

Constructors

OptDesc

Fields

optShort :: [Char]

Short option names, without the leading dash.

optLong :: [String]

Long option names, without the leading dashes.

optArgDesc :: ArgDesc opts err

Whether this option takes any arguments and, if so, how many.

data CmdDesc cmd opts posargs err

Describes a command. For instance, from the command line pantry find -i name Chex, the command is find.

Constructors

CmdDesc

Fields

cmdName :: String

The command name, such as find.

cmdF :: String -> cmd

Takes the full name of the command, returns something to indicate which command was called.

cmdOpts :: [OptDesc opts err]

A list of all the options that the command takes.

cmdPos :: PosDesc opts posargs err

How to parse positional arguments for this command.

parseGlobalsCmds

Arguments

:: (ParseErr err, Error err)

=> [OptDesc opts err]
Global options. For example, in the command pantry -i find name pretzels, -i is a global option because it appears before the command.

-> [CmdDesc cmd opts posargs err]
All command descriptions

-> opts
Default options. When the command line is parsed, each option receives the options that have already been parsed. The first option will receive this item.

-> [String]
What to parse

-> Either err (cmd, opts, posargs)
Left if an error occurred; Right if everything succeeded. The tuple has the command that was seen, the final state of the options item, and any positional arguments.

Parse a command line that features both global options and commands. This is the most heavy-duty parser. Do not include the program name as the first string to parse (consistent with System.Environment.getArgs).

parseCmds

Arguments

:: (ParseErr err, Error err)

=> [CmdDesc cmd opts posargs err]
All command descriptions

-> opts
Default options. When the command line is parsed, each option receives the options that have already been parsed. The first option will receive this item.

-> [String]
What to parse

-> Either err (cmd, opts, posargs)
Left if an error occurred; Right if everything succeeded. The tuple has the command that was seen, the final state of the options item, and any positional arguments.

Parses commands, but no global options.

parseOptsArgs

Arguments

:: (ParseErr err, Error err)

=> [OptDesc opts err]

-> opts

-> ([(opts, String)] -> Either err posargs)
How to parse a list of (opts, String) pairs. This single function will be called after all positional arguments have been processed. The opts element of the pair specifies the options that were in force when the positional argument was encountered on the command line.

-> [String]
What to parse. Do not include the program name in the list of options to parse (this is consistent with what is returned by System.Environment.getArgs).

-> Either err (opts, posargs)

Parses a command line that has options and positional arguments. This is the simplest parser and the one that most resembles many other parsers, such as System.Console.Getopt.